[pnnl/SHAD#115] Move shad::impl::array with shad::array.
I have if-zeroed out some tests from the array_test.  I am planning to reuse
some of that code and in the following patches I will move it in the new test.
